[editor/canvas state] Update and factor out copy/paste tests.
